Fix various bugs reported though the Android Market

+++ library/src/fm/last/api/LastFmServer.java
@@ -376,8 +376,8 @@ public interface LastFmServer {
 	
 	public Serializable[] multiSearch(String query) throws IOException, WSError;
 	
-	public void scrobbleTrack(String artist, String track, String album, long timestamp, int duration, String context, String streamid, String sk) throws IOException, WSError;
+	public void scrobbleTrack(String artist, String track, String album, long timestamp, long duration, String context, String streamid, String sk) throws IOException, WSError;
 	
-	public void updateNowPlaying(String artist, String track, String album, int duration, String context, String sk) throws IOException, WSError;
-	public void removeNowPlaying(String artist, String track, String album, int duration, String context, String sk) throws IOException, WSError;
+	public void updateNowPlaying(String artist, String track, String album, long duration, String context, String sk) throws IOException, WSError;
+	public void removeNowPlaying(String artist, String track, String album, long duration, String context, String sk) throws IOException, WSError;
 }
